pyproject.toml
[project]
name = "ward"
version = "0.1.0"
description = "Financial market data analysis tool with Web UI"
requires-python = ">=3.12"
dependencies = [
    "akshare>=1.18.0",
    "anthropic>=0.90.0",
    "chromadb>=1.5.0",
    "click>=8.3.0",
    "fastapi>=0.135.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    "uvicorn>=0.42.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ward = "ward.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
